+++ Config/DefaultEngine.ini
@@ -1,7 +1,7 @@
 [/Script/EngineSettings.GameMapsSettings]
-GameDefaultMap=/Game/FirstPerson/Maps/SMMap.SMMap
-EditorStartupMap=/Game/FirstPerson/Maps/SMMap.SMMap
-GlobalDefaultGameMode=/Game/FirstPerson/Blueprints/BP_FirstPersonGameMode.BP_FirstPersonGameMode_C
+GameDefaultMap=/Game/NotSharedResources/Prototype/ForTestMap.ForTestMap
+EditorStartupMap=/Game/NotSharedResources/Prototype/ForTestMap.ForTestMap
+GlobalDefaultGameMode=/Game/BP_Dungeon_ArmoryGameMode.BP_Dungeon_ArmoryGameMode_C
 GameInstanceClass=/Game/GI.GI_C
 
 [/Script/Engine.RendererSettings]
