+++ Config/DefaultGame.ini
@@ -1,3 +1,5 @@
 [/Script/EngineSettings.GeneralProjectSettings]
 ProjectID=124011CC436F7457BCCACB990DF71463
-ProjectName=Third Person Game Template
+ProjectName=Dungeon_Armory
+ProjectVersion=0.1.0
+
