+++ Content/Characters/Mob/Slime/Animation/Slime_Attack01_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a9144ba670ba5334079b089fa9b02ebe17120434279a477e2d4f1d23e974c546
+size 41957
