+++ Content/Characters/Mob/Slime/Animation/Slime_Attack02_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8b6a8df6c836cefff9b3f2c28d32634a044b63199061d70abbc1b9261797419c
+size 46287
