+++ Content/Characters/Mob/Slime/Animation/Slime_Die_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1225227c02edc7fab7b504865f4a94d41ecb3b237cd6656cb8488fa23d541442
+size 66616
