+++ Content/Characters/Mob/Slime/Animation/Slime_GetHit_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4582454de577675319f8b85fe76bec2af73329162fb1e8821d9f4a156312c003
+size 45939
