+++ Content/Characters/Mob/Slime/Animation/Slime_IdleBattle_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6b2229cea233a8fd7398f491e1e7adc51abf519913655a2d7bad6c18630ef967
+size 37502
