+++ Content/Characters/Mob/Slime/Animation/Slime_IdleNormal_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:02dd491baa0eedb9c819da54734a898d1870459a03f5189d2b6653a0f2c1612f
+size 45764
