+++ Content/Characters/Mob/Slime/Animation/Slime_Run_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:760e6cb38cb58ee481c772004bed450c7bf286b81df7a1b91e183f089e5f90b9
+size 43635
