+++ Content/Characters/Mob/Slime/Animation/Slime_SenseSomethingStart_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c9aa333362c9505043263be8a85ef471e8971359cfe0455a31f5d5f2d367c002
+size 72589
