+++ Content/Characters/Mob/Slime/Animation/Slime_Taunt_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a6839bd3b79600d6cc750fbe829919a983acc0f64966aad6723cb0748eed45f4
+size 39645
