+++ Content/Characters/Mob/Slime/Animation/Slime_Victory_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e287197a2763dcd9ef3c480e77037dd06b72d0c8b2ad133b36d0b74db4a8f5f5
+size 57242
