+++ Content/Characters/Mob/Slime/Animation/Slime_WalkBack_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5ba16c0f390637a47cde1235b43071d7c8f119460bc34a8baea217e07070c339
+size 46935
