+++ Content/Characters/Mob/Slime/Animation/Slime_WalkLeft_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2dd68ab57f1b40a5d58a32d405b424fd80348a49eb0dbeef919c3a462fdcb478
+size 50958
