+++ Content/Characters/Mob/Slime/Animation/Slime_WalkRight_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:616a92a655887930760e18bc327f14dfbea4b9f3ebfd05d83cd4303af4cee867
+size 50903
