+++ Content/Characters/Mob/Slime/Animation/Slime_Walk_ANIM.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c13d1b6792b6f64e79e52a13edffbb4a79d2f67a0dff231d2a095c9dd8f6f6b5
+size 45014
