+++ Content/Characters/Mob/Slime/Blueprint/BP_Slime.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0a2cc032b84dcb1e27eb2a710cb64ab3d1c95180ddc90d47125ac8c97b6139e2
+size 33032
