+++ Source/Dungeon_Armory/Private/Characters/Core/AIControllerBase.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "Characters/NPC/NPCAIControllerBase.h"
+#include "Characters/Core/AIControllerBase.h"
 #include "Characters/NPC/NPCBase.h"
 
 #include "Perception/AIPerceptionComponent.h"
@@ -15,7 +15,7 @@
 
 #include "AI/Team/TeamComponent.h"
 
-ANPCAIControllerBase::ANPCAIControllerBase()
+AAIControllerBase::AAIControllerBase()
 {
     PrimaryActorTick.bCanEverTick = false;
 
@@ -36,19 +36,19 @@ ANPCAIControllerBase::ANPCAIControllerBase()
 
     AIPerceptionComponent->ConfigureSense(*SightConfig);
     AIPerceptionComponent->SetDominantSense(SightConfig->GetSenseImplementation());
-    AIPerceptionComponent->OnTargetPerceptionUpdated.AddDynamic(this, &ANPCAIControllerBase::OnTargetPerceived);
+    AIPerceptionComponent->OnTargetPerceptionUpdated.AddDynamic(this, &AAIControllerBase::OnTargetPerceived);
 
     // 팀 아이디 초기화
 	TeamId = FGenericTeamId::NoTeam;
 }
 
-void ANPCAIControllerBase::BeginPlay()
+void AAIControllerBase::BeginPlay()
 {
     Super::BeginPlay();
 
 }
 
-void ANPCAIControllerBase::OnPossess(APawn* InPawn)
+void AAIControllerBase::OnPossess(APawn* InPawn)
 {
     Super::OnPossess(InPawn);
 
@@ -60,37 +60,15 @@ void ANPCAIControllerBase::OnPossess(APawn* InPawn)
             TeamComponent = TeamCmp; // 팀 정보를 가져와 저장
         }
     }
-
-    // 비헤이비어 트리 실행
-    if (BehaviorTree && BehaviorTree->BlackboardAsset)
-    {
-        // 블랙보드 설정 (AIController에서 관리됨)
-        UseBlackboard(BehaviorTree->BlackboardAsset, BlackboardComponent);
-
-        // 블랙보드 변수 초기화
-        if (BlackboardComponent)
-        {
-            BlackboardComponent->SetValueAsFloat("MovementSpeed", 500.0f);
-
-            BlackboardComponent->SetValueAsVector("HomeLocation", GetPawn()->GetActorLocation());
-            SetNPCState(ENPCStates::Stay);
-
-            // 비헤이비어 트리 실행 (이 시점에서 BehaviorTreeComponent가 자동으로 생성됨)
-            RunBehaviorTree(BehaviorTree);
-
-            // 내부적으로 생성된 BehaviorTreeComponent를 가져와서 멤버 변수에 할당
-            BehaviorTreeComponent = FindComponentByClass<UBehaviorTreeComponent>();
-        }
-    }
 }
 
-void ANPCAIControllerBase::Tick(float DeltaTime)
+void AAIControllerBase::Tick(float DeltaTime)
 {
     Super::Tick(DeltaTime);
 
 }
 
-ETeamAttitude::Type ANPCAIControllerBase::GetTeamAttitudeTowards(const AActor& Other) const
+ETeamAttitude::Type AAIControllerBase::GetTeamAttitudeTowards(const AActor& Other) const
 {
     if (TeamComponent)
     {
@@ -100,7 +78,7 @@ ETeamAttitude::Type ANPCAIControllerBase::GetTeamAttitudeTowards(const AActor& O
 	return ETeamAttitude::Neutral;
 }
 
-FGenericTeamId ANPCAIControllerBase::GetGenericTeamId() const
+FGenericTeamId AAIControllerBase::GetGenericTeamId() const
 {
     if (TeamComponent)
     {
@@ -111,7 +89,7 @@ FGenericTeamId ANPCAIControllerBase::GetGenericTeamId() const
 }
 
 // 감지 이벤트 처리
-void ANPCAIControllerBase::OnTargetPerceived(AActor* Actor, FAIStimulus Stimulus)
+void AAIControllerBase::OnTargetPerceived(AActor* Actor, FAIStimulus Stimulus)
 {
     if (BlackboardComponent == nullptr)
     {
@@ -128,23 +106,4 @@ void ANPCAIControllerBase::OnTargetPerceived(AActor* Actor, FAIStimulus Stimulus
     }
 }
 
-ENPCStates ANPCAIControllerBase::GetNPCState() const
-{
-    if (BlackboardComponent == nullptr)
-    {
-        return ENPCStates::None;
-    }
 
-    int8 CurrNPCState = BlackboardComponent->GetValueAsEnum(BBKey_NPCState);
-    return static_cast<ENPCStates>(CurrNPCState);
-}
-
-void ANPCAIControllerBase::SetNPCState(ENPCStates NewNPCState)
-{
-    if (BlackboardComponent == nullptr)
-    {
-        return;
-    }
-
-    BlackboardComponent->SetValueAsEnum(BBKey_NPCState, static_cast<uint8>(NewNPCState));
-}
