+++ Source/Dungeon_Armory/Private/Characters/Mob/AIController/MobAIController.cpp
@@ -1,0 +1,41 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Characters/Mob/AIController/MobAIController.h"
+
+#include "BehaviorTree/BehaviorTree.h"
+#include "BehaviorTree/BehaviorTreeComponent.h"
+#include "BehaviorTree/BlackboardComponent.h"
+
+AMobAIController::AMobAIController()
+{
+}
+
+void AMobAIController::BeginPlay()
+{
+}
+
+void AMobAIController::OnPossess(APawn* InPawn)
+{
+    // 팀 컴포넌트 설정
+    Super::OnPossess(InPawn);
+
+    // 비헤이비어 트리 실행
+    if (BehaviorTree && BehaviorTree->BlackboardAsset)
+    {
+        // 블랙보드 설정 (AIController에서 관리됨)
+        UseBlackboard(BehaviorTree->BlackboardAsset, BlackboardComponent);
+
+        // 블랙보드 변수 초기화
+        if (BlackboardComponent)
+        {
+            BlackboardComponent->SetValueAsVector("HomeLocation", GetPawn()->GetActorLocation());
+
+            // 비헤이비어 트리 실행 (이 시점에서 BehaviorTreeComponent가 자동으로 생성됨)
+            RunBehaviorTree(BehaviorTree);
+
+            // 내부적으로 생성된 BehaviorTreeComponent를 가져와서 멤버 변수에 할당
+            BehaviorTreeComponent = FindComponentByClass<UBehaviorTreeComponent>();
+        }
+    }
+}
