+++ Source/Dungeon_Armory/Private/Characters/Mob/Tasks/BTTask_Patrol.cpp
@@ -1,0 +1,5 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Characters/Mob/Tasks/BTTask_Patrol.h"
+
