+++ Source/Dungeon_Armory/Public/Characters/Mob/AIController/MobAIController.h
@@ -1,0 +1,25 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Characters/Core/AIControllerBase.h"
+#include "MobAIController.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DUNGEON_ARMORY_API AMobAIController : public AAIControllerBase
+{
+	GENERATED_BODY()
+	
+/***** Unreal *****/
+public:
+	AMobAIController();
+
+protected:
+	virtual void BeginPlay() override;
+	virtual void OnPossess(APawn* InPawn) override;
+
+};
