+++ Source/Dungeon_Armory/Public/Characters/Mob/Tasks/BTTask_Patrol.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "BehaviorTree/Tasks/BTTask_BlackboardBase.h"
+#include "BTTask_Patrol.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DUNGEON_ARMORY_API UBTTask_Patrol : public UBTTask_BlackboardBase
+{
+	GENERATED_BODY()
+	
+};
